+++ services/preview/confluent/mgmt/2021-03-01-preview/confluent/models.go
@@ -492,49 +492,6 @@ type OperationResult struct {
 	IsDataAction *bool `json:"isDataAction,omitempty"`
 }
 
-// OrganizationCreateFuture an abstraction for monitoring and retrieving the results of a long-running
-// operation.
-type OrganizationCreateFuture struct {
-	azure.FutureAPI
-	// Result returns the result of the asynchronous operation.
-	// If the operation has not completed it will return an error.
-	Result func(OrganizationClient) (OrganizationResource, error)
-}
-
-// UnmarshalJSON is the custom unmarshaller for CreateFuture.
-func (future *OrganizationCreateFuture) UnmarshalJSON(body []byte) error {
-	var azFuture azure.Future
-	if err := json.Unmarshal(body, &azFuture); err != nil {
-		return err
-	}
-	future.FutureAPI = &azFuture
-	future.Result = future.result
-	return nil
-}
-
-// result is the default implementation for OrganizationCreateFuture.Result.
-func (future *OrganizationCreateFuture) result(client OrganizationClient) (or OrganizationResource, err error) {
-	var done bool
-	done, err = future.DoneWithContext(context.Background(), client)
-	if err != nil {
-		err = autorest.NewErrorWithError(err, "confluent.OrganizationCreateFuture", "Result", future.Response(), "Polling failure")
-		return
-	}
-	if !done {
-		or.Response.Response = future.Response()
-		err = azure.NewAsyncOpIncompleteError("confluent.OrganizationCreateFuture")
-		return
-	}
-	sender := autorest.DecorateSender(client, autorest.DoRetryForStatusCodes(client.RetryAttempts, client.RetryDuration, autorest.StatusCodesForRetry...))
-	if or.Response.Response, err = future.GetResult(sender); err == nil && or.Response.Response.StatusCode != http.StatusNoContent {
-		or, err = client.CreateResponder(or.Response.Response)
-		if err != nil {
-			err = autorest.NewErrorWithError(err, "confluent.OrganizationCreateFuture", "Result", or.Response.Response, "Failure responding to request")
-		}
-	}
-	return
-}
-
 // OrganizationDeleteFuture an abstraction for monitoring and retrieving the results of a long-running
 // operation.
 type OrganizationDeleteFuture struct {