+++ services/preview/confluent/mgmt/2021-03-01-preview/confluent/CHANGELOG.md
@@ -1,2 +1,17 @@
-# Change History
+# Unreleased
 
+## Breaking Changes
+
+### Removed Funcs
+
+1. *OrganizationCreateFuture.UnmarshalJSON([]byte) error
+1. OrganizationClient.Create(context.Context, string, string, *OrganizationResource) (OrganizationCreateFuture, error)
+1. OrganizationClient.CreatePreparer(context.Context, string, string, *OrganizationResource) (*http.Request, error)
+1. OrganizationClient.CreateResponder(*http.Response) (OrganizationResource, error)
+1. OrganizationClient.CreateSender(*http.Request) (OrganizationCreateFuture, error)
+
+### Struct Changes
+
+#### Removed Structs
+
+1. OrganizationCreateFuture